+++ packages/MagicCall/example/eg_magiccall_v1.jl
@@ -17,8 +17,8 @@ pedinfo = string(dataid, "_magicsimulate_ped.csv")
 
 @time magiccall(genofile,pedinfo;     
     isfounderinbred,
-    israwcall = true, 
-    # likeparameters = LikeParam(0.005, 0.005, 0.0, nothing, nothing, nothing, 0.0),
+    # israwcall = true, 
+    # likeparam = LikeParam(0.005, 0.005, 0.0, 0.001, nothing, nothing, 0.0),
     # model = "depmodel", 
     # isparallel = false, 
 )
@@ -30,10 +30,8 @@ magicgeno = formmagicgeno(calledgenofile,pedinfo;
 );
 
 
-
 truefile = "sim_magicsimulate_truegeno.csv.gz"
 truegeno = formmagicgeno(truefile,pedinfo;isfounderinbred);
-
 acc = magicaccuracy!(truegeno, magicgeno; isfounderinbred)
 println(acc)
 
@@ -67,23 +65,9 @@ display(fig)
 
 # clear up
 
-cd(@__DIR__)
-dataid = "sim"
-rm.(filter(x->occursin(dataid, x), readdir()))
-rm.(filter(x->occursin("outstem", x), readdir()))
+# cd(@__DIR__)
+# dataid = "sim"
+# rm.(filter(x->occursin(dataid, x), readdir()))
+# rm.(filter(x->occursin("outstem", x), readdir()))
 
-
-
-
-
-
-##########################3
-
-if !israwcall
-    msg = string("likeparameters = (foundererror, offspringerror, peroffspringerror, baseerror, allelicbias, allelicoverdispersion, allelicdropout)")
-    printconsole(logio,verbose,msg)    
-end
-
-priorlikeparameters = PriorLikeParam(baseerror=Beta(1,299))
-reset_priorlikeparameters(priorlikeparameters)