+++ packages/MagicImpute/src/magicimpute_fun.jl
@@ -59,7 +59,7 @@ function magicimpute(genofile::AbstractString,
     isbinning::Union{Nothing,Bool}=nothing,        
     bincm::Real=0.001, # successive markers with intermarker distance < bincm are binned    
     isinfererror::Union{Nothing, Bool} = true,               
-    tukeyfence::Real=1.5,         
+    tukeyfence::Real=2,         
     iscorrectfounder::Union{Nothing, Bool} = true,       
     phasealg::AbstractString="unphase", 
     isdelmarker::Bool= true,         
@@ -68,7 +68,7 @@ function magicimpute(genofile::AbstractString,
     isspacemarker::Bool= !isnothing(mapfile) || isordermarker || isphysmap,
     trimcm::Real=20,
 	trimfraction::Real=0.05,  #cM    
-    skeletonsize::Union{Nothing,Integer} = nothing,  
+    skeletonsize::Union{Nothing,Integer} = 100,  
     slidewin_neighbor::Union{Nothing,Integer} = 200,
     slidewin::Union{Nothing,Integer} = nothing,	            
     binriffle::Union{Nothing,Integer} = (!isnothing(mapfile) && isfounderinbred) ? -1 : nothing,  
@@ -77,7 +77,7 @@ function magicimpute(genofile::AbstractString,
     inittemperature::Real= isordermarker ? 2.0 : 0.0,
     coolrate::Real=0.8,
     minaccept::Real=0.15,
-    spacebyviterbi::Bool=false,     
+    spacebyviterbi::Bool=true,     
     isparallel::Bool=true,    
     isparallelfounder::Bool=true, 
     workdir::AbstractString=pwd(),
@@ -258,7 +258,7 @@ genotype imputation from magicgeno.
 
 `isinfererror::Bool = true`: if true, infer marker specific likelihood parameters that have values of nothing in likeparam. 
 
-`tukeyfence::Real=1.5`: tukey fence for detecting outlier error rates (including foundererror, offspringerror, baseerror, and allelicbias). 
+`tukeyfence::Real=2`: tukey fence for detecting outlier error rates (including foundererror, offspringerror, baseerror, and allelicbias). 
 
 `iscorrectfounder::Union{Nothing, Bool} = true`: if true, perform parental error correction.
 
@@ -279,7 +279,7 @@ genotype imputation from magicgeno.
 `trimcm::Real=20`: remove markers of each segment with distances to the flanking markers > trimcm.
   The number of markers of each segment must be less than 5% total number of markers.
 
-`skeletonsize::Union{Nothing,Integer} = nothing`: number of skeleton markers for piecewisely re-scaling inter-marker distances. 
+`skeletonsize::Union{Nothing,Integer} = 100`: number of skeleton markers for piecewisely re-scaling inter-marker distances. 
   If it is nothing, skeletonsize is set to the number of distint positions in the genetic map before re-scaling. 
 
 `slidewin_neighbor::Union{Nothing,Integer} = 200`: max sliding window size for neighbor-based marker order refinement. 
@@ -347,7 +347,7 @@ function magicimpute!(magicgeno::MagicGeno;
     nrepeatmax::Integer=6,     
     isinferjunc::Union{Nothing, Bool} = false,     
     isinfererror::Union{Nothing, Bool} = true,                
-    tukeyfence::Real=1.5,        
+    tukeyfence::Real=2,        
     iscorrectfounder::Union{Nothing, Bool} = true,     
     phasealg::AbstractString="unphase", 
     isdelmarker::Bool= true,
@@ -356,7 +356,7 @@ function magicimpute!(magicgeno::MagicGeno;
     isspacemarker::Bool= !isnothing(inputneighbor) || isordermarker,
     trimcm::Real=20,
 	trimfraction::Real=0.05,  #cM    
-    skeletonsize::Union{Nothing,Integer} = nothing,     
+    skeletonsize::Union{Nothing,Integer} = 100,     
     slidewin_neighbor::Union{Nothing,Integer} = 200,
     slidewin::Union{Nothing,Integer} = nothing,	        
     binriffle::Union{Nothing,Integer} = (!isnothing(inputneighbor) && isfounderinbred) ? -1 : nothing,  
@@ -365,7 +365,7 @@ function magicimpute!(magicgeno::MagicGeno;
     inittemperature::Real= isordermarker ? 2.0 : 0.0,
     coolrate::Real=0.8,
     minaccept::Real=0.15,
-    spacebyviterbi::Bool=false,     
+    spacebyviterbi::Bool=true,     
     isparallel::Bool=true,
     isparallelfounder::Bool=true, 
     workdir::AbstractString=pwd(),